requirements.txt
transformers>=4.20.0
torch>=1.10.0
torchvision>=0.11.0
pandas>=1.3.0
numpy>=1.19.3,<2.0.0
Pillow>=8.0.0,<11.0.0
pdf2image>=1.16.0
openpyxl>=3.0.7
pytesseract>=0.3.8
tqdm>=4.62.0
opencv-python>=4.5.0
scikit-image>=0.18.0